Create simon_levels
This memory game is what is often called a 'Simon says' in which the program generates a random sequence of 4 colours (each colour also has it's own tone) which you have to recreate in a limited time. A wrong sequence or a time-out gives a low buzz and a red light, a right sequence gives a bright sound and flashing lights. After each success, the sequence gets longer by 1, so that a difficulty level is added. Sorry, serial messages and some of the variables are in Dutch, but those are easily edited.